requirements.txt
attrs==21.4.0
Django==4.0.5
django-extensions==3.1.5
djangorestframework==3.13.1
Faker==13.14.0
importlib-metadata==4.11.4
more-itertools==8.13.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3
wcwidth==0.2.5
zipp==3.8.0
